Started to refactor opcodes non dependent on operating modes

+++ src/main/java/com/github/smeny/jpc/emulator/execution/opcodes/rm/add_Eb_Gb_mem.java
@@ -37,32 +37,32 @@ import com.github.smeny.jpc.emulator.processor.Processor;
 import static com.github.smeny.jpc.emulator.processor.Processor.Reg;
 
 public class add_Eb_Gb_mem extends Executable {
-  final Pointer op1;
-  final int op2Index;
+    final Pointer op1;
+    final int op2Index;
 
-  public add_Eb_Gb_mem(int blockStart, int eip, int prefices, PeekableInputStream input) {
-    super(blockStart, eip);
-    int modrm = input.readU8();
-    op1 = Modrm.getPointer(prefices, modrm, input);
-    op2Index = Modrm.Gb(modrm);
-  }
+    public add_Eb_Gb_mem(int blockStart, int eip, int prefices, PeekableInputStream input) {
+        super(blockStart, eip);
+        int modrm = input.readU8();
+        op1 = Modrm.getPointer(prefices, modrm, input);
+        op2Index = Modrm.Gb(modrm);
+    }
 
-  public Branch execute(Processor cpu) {
-    Reg op2 = cpu.regs[op2Index];
-    cpu.flagOp1 = op1.get8(cpu);
-    cpu.flagOp2 = (byte) op2.get8();
-    cpu.flagResult = (byte) (cpu.flagOp1 + cpu.flagOp2);
-    op1.set8(cpu, (byte) cpu.flagResult);
-    cpu.flagIns = UCodes.ADD8;
-    cpu.flagStatus = OSZAPC;
-    return Branch.None;
-  }
+    public Branch execute(Processor cpu) {
+        Reg op2 = cpu.regs[op2Index];
+        cpu.flagOp1 = op1.get8(cpu);
+        cpu.flagOp2 = op2.get8();
+        cpu.flagResult = (byte) (cpu.flagOp1 + cpu.flagOp2);
+        op1.set8(cpu, (byte) cpu.flagResult);
+        cpu.flagIns = UCodes.ADD8;
+        cpu.flagStatus = OSZAPC;
+        return Branch.None;
+    }
 
-  public boolean isBranch() {
-    return false;
-  }
+    public boolean isBranch() {
+        return false;
+    }
 
-  public String toString() {
-    return this.getClass().getName();
-  }
+    public String toString() {
+        return this.getClass().getName();
+    }
 }
